Fix Sprite Assembly
This optimization doesn't account for the width of the assembled image
possibly being too small to fit the whole source image width. Since it
meant copying more image data than we needed, I'm no longer convinced
that it actually is an optimization over a simple cut (as this patch
implements).

Change-Id: I3fe502a75979e972f62d9dbf26f1ffec9069a102
Reviewed-by: Yann Bodson <yann.bodson@nokia.com>