Optimise QQuickItem::stackAfter/Before() for the common case.
These functions are most often used to position items just given a new
parent, meaning one item is very likely to be at the end of the
child list, and when appending multiple items the other is also likely
to be near the end.  When this is true we can save a lot of time
spent iterating to the end of the list by doing a reverse search.

Change-Id: If16d6d19689a6140d3d9a873857776cf79666910
Reviewed-by: Martin Jones <martin.jones@nokia.com>