Keep polishing until there is nothing left to polish.
Some items will trigger changes in updatePolish() which
again trigger property changes which needs polish to
be executed again. If we do not polish until the set
is fully cleared, we will end up with some items being
in an unpolished state.

To avoid polishing from going on indefinitely, we have a
countdown that spits out a warning if polishing goes
on forever.

Change-Id: I4da6e884ca0d52b9b2701082ecf13e8c65508524
Reviewed-by: Aaron Kennedy <aaron.kennedy@nokia.com>