Remove non-templated qobject module api registration function
Previously, both templated and non-templated qobject module api
registration functions were provided as public API.  The non-templated
version is identical to the templated version, except that bindings
which access properties of module apis registered with that function
cannot be optimized with v4.

This commit removes the non-templated version of the function.

Change-Id: Ic417aa241f1172bf6d055bdc7f528d7e9304097b
Reviewed-by: Martin Jones <martin.jones@nokia.com>