Make qqmlmoduleplugin tests pass in shadow builds.
The plugin binary and qml files for a module need to be in the same
directory.  This was solved for source builds because the files were
already located in the import path, but with shadow builds the files
were split between the build and source trees.  To solve this we copy
the files to the import path when doing a build.  So no files are
copied on top of themselves all mixed module files have been relocated
to their module source directory.

Change-Id: I238af998a0f766e67ed6d0023e5ab4c2a4ea67af
Reviewed-by: Chris Adams <christopher.adams@nokia.com>