Evaluate bindings more intelligently during construction
Instead of just evaluating bindings in a fixed order, and possibly
having to evaluate a single binding multiple times, prior to reading
a property, we check if there are any bindings "pending" on it and
evaluate them then.  A pending binding is one that has been assigned
to the property, but not yet evaluated.

To minimize side effects we only do this for "safe" bindings.  A
safe binding is one that has no side effects, which we currently
define as not calling functions or otherwise assigning values during
its evaluation.  This isn't an entirely foolproof way to ensure that
the evaluation has no side effects, but it should be good enough.

Change-Id: I98aa76a95719e5d182e8941738d64f8d409f404a
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>