Fix crash in lazy meta object generation.
Inside the property cache, override data can only handle overrides
pointing to either a method or a property, but not to a signal
handler.  This is a bug in itself, but has never come up before due
to no one following override data on methods.  When this
was changed by d2e557c2c2d7fcf3bf7c1676df3902e115986dc2, this bug
was exposed.

This change doesn't actually fix the underlying problem, but it does
restore exactly the same behavior we had previously.  The complete
fix will come in a later change.

Change-Id: I6a890e6ca1e40735da8158b21dfe38dc88091081
Reviewed-by: Martin Jones <martin.jones@nokia.com>