Fix handling of a direct child stealing focus from a parent-less item.
When an item gains focus it will remove focus from the previous item
that held focus.  Normally this is the sub-focus item of an ancestor
item, however if the item with focus is at the root of the tree there
will be no sub-focus item and the focus of root item needs to be cleared
instead.  Likewise if an item doesn't have a parent, focus still needs
to be removed from any sub focus item.

Change-Id: I0f192692e21e288c87ba0104f3d597363190f87c
Reviewed-by: Martin Jones <martin.jones@nokia.com>