Emit itemChanged when Loader fails to load a source.
As per the source comment, it is not actually tracked whether item was
null before. This is consistent with the behaviour in loadFromSource,
where setting a source to null always emits item changed.

This is reasonable behaviour, because it only occurs in unexpected error
states where the program logic likely expects a changed signal anyways.

Change-Id: I160ab04dc1b5c578ee508f4adbca65e8da5b16be
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>