Fix locale test on Linux and skip it on Windows.
The timezone that V8 internally uses was not being updated at the
end of one test, causing failures in subsequent tests. This was
only showing up on Windows because the test was being skipped on
Linux due to a typo. Skip the test on Windows due to complexity
of changing the timezone (added comments in code with details).

Change-Id: I47c8542111e8ddfbdeff39815c50d98570b0c6c2
Reviewed-by: Matthew Vogt <matthew.vogt@nokia.com>