pressed and pressedButtons don't work when multiple buttons are pressed
The pressed property would react to any mouse event, regardless of
whether it was in acceptedButtons, or there were other buttons still
pressed.  It will now remain true while any button in acceptedButtons
is pressed.

The pressedButtons property could contain buttons not in the
acceptedButtons mask.  It will now only contain buttons that are in the
acceptedButtons mask.

Task-number: QTBUG-26458
Change-Id: I6b25cc36a58c113de062d530761dc179d7ef4a5a
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>