Fix inconsistent playing / paused property setting  and signals
Playing should only indicate false if it has been set as such
in code, or in the event of a valid image but non-movie being
used.
Noted the edge case in documentation.
Signals were not being emitted for playing/paused in the cases
where an image was not loaded. Paused property was also being
incorrectly assigned, which supressed pause behavior.

Task-number: QTBUG-26478
Change-Id: I1aaab9b7c495811114b30b491a900804a75d590f
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>