Fix resolution of cursor when items are overlapping.
If MouseArea with cursorShapes are overlapping then cursor shape
of the foremost item under the mouse cursor should be shown, but because
the hover events are delivered to the foremost items first the opposite
is occuring.  This makes QQuickWindow responsible for correctly setting
its own cursor instead of relying on items to work it out amongst
themselves.

Change-Id: Iedf144c583dfa3d1ff441e19db9601b5e171902a
Reviewed-by: Martin Jones <martin.jones@nokia.com>